requirements.txt
numpy~=1.18.4
pandas~=1.0.3
matplotlib
biopython
